+++ cauliflower/driver_rp2.py
@@ -1,7 +1,7 @@
 import time
 
-from cauliflower.log import error, trace, debug, info, LogLevel
-from cauliflower.nand import NandConfig, NandCmd, NandStatus
+from .log import error, trace, debug, info, LogLevel
+from .nand import NandConfig, NandCmd, NandStatus
 
 from machine import Pin
